requirements.txt
# requirements.txt
Flask
Flask-Mail
python-dotenv